requirements.txt
flask==2.3.2
psycopg2-binary==2.9.7
gunicorn==21.2.0
Flask-Cors==3.1.1